refactor results into component
